refactor(core): extract indefinite article logic into its own module
Move InitialSound enum, starts_with_vowel(), to_lower_word(), and
is_likely_acronym() from linting/an_a.rs into a dedicated
indefinite_article module. This makes the a/an heuristics reusable
by future linters that need to determine the correct article.

Fixes #2979
