+++ harper-core/src/lib.rs
@@ -12,6 +12,7 @@ mod edit_distance;
 pub mod expr;
 mod fat_token;
 mod ignored_lints;
+mod indefinite_article;
 mod irregular_nouns;
 mod irregular_verbs;
 pub mod language_detection;
@@ -50,6 +51,7 @@ pub use dict_word_metadata_orthography::{OrthFlags, Orthography};
 pub use document::Document;
 pub use fat_token::{FatStringToken, FatToken};
 pub use ignored_lints::{IgnoredLints, LintContext};
+pub use indefinite_article::{InitialSound, starts_with_vowel};
 pub use irregular_nouns::IrregularNouns;
 pub use irregular_verbs::IrregularVerbs;
 use linting::Lint;
